TrainCheckpoint: 'push_to_hub' keyword added and saving 'model-checkpoint.pth'
`push_to_hub:
Other checkpoint class has it and Trainer passes a value for it.
Thus, the trainer could not be used with this checkpoint class.

`model-checkpoint.pth`
Somehow this was missing from the class.
Maybe got deleted during 'push_to_hub' implementation?
Could not load trained models without it.